Separate piece inference
